Un poco de limpieza de código, problema muy extraño con unordered_map.

+++ src/vehicle.cpp
@@ -6,21 +6,21 @@ Vehicle::Vehicle(unsigned short pChromosome)
 	this->fitnessScore = 0.0;
 }
 
-short Vehicle::getTreadId()
+int Vehicle::getTreadId()
 {
 	unsigned char c = 0;
 	c = c | this->chromosome;
 	double num = c;
-	return 11 - ceil(num/CHROMOSOME_RANGE_SIZE);
+	return (int) (11 - ceil(num/CHROMOSOME_RANGE_SIZE));
 }
 
-short Vehicle::getTorqueId()
+int Vehicle::getTorqueId()
 {
 	this->chromosome = this->chromosome >> 8;
 	unsigned char c = 0;
 	c = c | this->chromosome;
 	double num = c;
-	return ceil(num/CHROMOSOME_RANGE_SIZE);
+	return (int) ceil(num/CHROMOSOME_RANGE_SIZE);
 }
 
 void Vehicle::setFitnessScore(double pScore)
